# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Carteiras de jogadores e operações de provedores de jogos: apostas, prêmios, derrotas, estornos e reversões.",
        "title": "pedro-test",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/providers/{providerId}/wagering/transactions/{externalTransactionId}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "O providerId do caminho precisa ser o do token: outro é 403, decidido antes de tocar em qualquer dado.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wagering"
                ],
                "summary": "Lê uma transação pelo id do provedor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id do provedor, o mesmo do token",
                        "name": "providerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Id que o provedor deu à transação",
                        "name": "externalTransactionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wagering.TransactionDetail"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        },
        "/wagering/transactions": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aplica BET, WIN, LOSS, REFUND ou ROLLBACK a uma carteira, uma única vez, por mais que seja reenviada. O header Idempotency-Key é obrigatório e nunca é substituído pelo servidor. Mesma chave e mesmo conteúdo devolvem o resultado gravado, com o saldo observado no processamento original (200 ou 422, com idempotentReplay true). Mesma chave com outro conteúdo, ou a mesma operação sob outra chave, é 409. O providerId do corpo precisa ser o do token.\nDistinguível pelo status: 200 processada, 202 aguardando referência, 400 entrada inválida (nada gravado), 409 conflito de idempotência, 422 rejeição de negócio (com failureCode, gravada e reproduzível), 503 indisponibilidade transitória (tente de novo).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wagering"
                ],
                "summary": "Envia uma operação de um provedor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Chave escolhida pelo provedor, por exemplo provider-a:transaction-123",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "A operação",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/wagering.SubmitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wagering.TransactionResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/wagering.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/wagering.TransactionResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        },
        "/wagering/transactions/{transactionId}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Permite acompanhar uma pendência (tentativas, próxima olhada e expiração) e consultar o código de uma rejeição ou falha. Uma transação de outro provedor responde 404, exatamente como uma que não existe: perguntar não revela de quem é.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wagering"
                ],
                "summary": "Lê uma transação pelo id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id da transação",
                        "name": "transactionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/wagering.TransactionDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/structs.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "structs.APIError": {
            "type": "object",
            "properties": {
                "failureCode": {
                    "description": "FailureCode is the stable, documented reason of a business rejection or of an invalid\noperation. It is what a provider reads to tell an input it can correct from a definitive\noutcome, and it is absent on errors that are not about the operation itself.",
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "message": {
                    "type": "string",
                    "example": "invalid token"
                }
            }
        },
        "structs.MoneyDTO": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "25.00"
                },
                "currency": {
                    "type": "string",
                    "example": "BRL"
                }
            }
        },
...
        "wagering.SubmitRequest": {
            "type": "object",
            "properties": {
                "externalTransactionId": {
                    "type": "string",
                    "example": "transaction-123"
                },
                "gameId": {
                    "type": "string",
                    "example": "fortune-chimp"
                },
                "kind": {
                    "type": "string",
                    "enum": [
                        "BET",
                        "WIN",
                        "LOSS",
                        "REFUND",
                        "ROLLBACK"
                    ],
                    "example": "BET"
                },
                "money": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "playerId": {
                    "type": "string",
                    "example": "0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1"
                },
                "providerId": {
                    "type": "string",
                    "example": "provider-a"
                },
                "referenceExternalTransactionId": {
                    "type": "string",
                    "example": "transaction-122"
                },
                "roundId": {
                    "type": "string",
                    "example": "round-987"
                },
                "walletId": {
                    "type": "string",
                    "example": "0192f291-27dd-7d3f-8071-5f8685deef37"
                }
            }
        },
        "wagering.TransactionDetail": {
            "type": "object",
            "properties": {
                "balance": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2026-09-08T12:00:00.000Z"
                },
                "expiresAt": {
                    "type": "string",
                    "example": "2026-09-09T12:00:00.000Z"
                },
                "externalTransactionId": {
                    "type": "string",
                    "example": "transaction-123"
                },
                "failureCode": {
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "gameId": {
                    "type": "string",
                    "example": "fortune-chimp"
                },
                "kind": {
                    "type": "string",
                    "enum": [
                        "BET",
                        "WIN",
                        "LOSS",
                        "REFUND",
                        "ROLLBACK"
                    ],
                    "example": "BET"
                },
                "money": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "nextAttemptAt": {
                    "type": "string",
                    "example": "2026-09-08T12:00:05.000Z"
                },
                "playerId": {
                    "type": "string",
                    "example": "0192f28f-5dc0-7d58-bdb2-814ad6a0f4a1"
                },
                "providerId": {
                    "type": "string",
                    "example": "provider-a"
                },
                "referenceAttempts": {
                    "type": "integer",
                    "example": 3
                },
                "referenceExternalTransactionId": {
                    "type": "string",
                    "example": "transaction-122"
                },
                "roundId": {
                    "type": "string",
                    "example": "round-987"
                },
                "settledAt": {
                    "type": "string",
                    "example": "2026-09-08T12:00:00.000Z"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "PENDING",
                        "PENDING_REFERENCE",
                        "PROCESSED",
                        "REJECTED",
                        "FAILED"
                    ],
                    "example": "PROCESSED"
                },
                "transactionId": {
                    "type": "string",
                    "example": "0192f298-345e-7e38-af88-e43f851a819d"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2026-09-08T12:00:00.000Z"
                },
                "walletId": {
                    "type": "string",
                    "example": "0192f291-27dd-7d3f-8071-5f8685deef37"
                }
            }
        },
        "wagering.TransactionResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "$ref": "#/definitions/structs.MoneyDTO"
                },
                "failureCode": {
                    "type": "string",
                    "example": "INSUFFICIENT_FUNDS"
                },
                "idempotentReplay": {
                    "type": "boolean",
                    "example": false
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "PROCESSED",
                        "REJECTED",
                        "PENDING_REFERENCE",
                        "FAILED"
                    ],
                    "example": "PROCESSED"
                },
                "transactionId": {
                    "type": "string",
                    "example": "0192f298-345e-7e38-af88-e43f851a819d"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Cole \"Bearer\" e um token. É o caminho das rotas que exigem um papel de serviço (`internal_service` ou `provider`): o Authorize acima faz password grant, e nenhum usuário do realm tem esses papéis. Pegue o token com `make token-internal` ou `make token-provider`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "http://localhost:8080/realms/pedro-test/protocol/openid-connect/token",
            "scopes": {
                "openid": "Acesso à API"
            }
        }
    }
}